refactor: 도메인 및 hooks 폴더 생성
